Don't impose kernel header dependency/warning on prebuilt kernels
These kernel header dependency and warning are great for when building
a kernel inline, they help avoid having modules left out of the kernel
due to a race between the kernel build and the modules. This is all stated
in the original commit.

The problem with this is that if you attempt to compile with a prebuilt kernel
for whatever reason, you'll be faced with warning because it can't tell the difference.

Change-Id: Id43518747d78632117ef7f798e6182b9f008673a